requirements.txt
streamlit
langchain
chromadb
sentence-transformers
pymupdf
langchain-text-splitters
python-dotenv
